+++ CHANGELOG.md
@@ -1,9 +1,20 @@
 # Change Log
 All notable changes to this project will be documented in this file. This change log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
-## 0.2.0 - 2019-11-14
+## [Unreleased]
 ### Added
-- def some operators
+- Added aggregate function
+- `skip` option to `find`
+
+### Changed
+- Better docs for cljdoc
+- Merged `find-maps` into `find`
+- Added ? suffix to boolean params
+- Renamed `find-one-as-map` to `find-one`
+
+## [0.2.0] - 2019-11-14
+### Added
+- expose operators
 
 ## 0.1.0 - 2019-11-14
 ### Added